+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+### Changed
+
+- When a connection drop the inflight manager will now re-transmit
+  packages with quality of service that was inflight when the regain a
+  connection.
+
+- Improved the logic for handling incoming publishes with Quality of
+  Service 2 (exactly once delivery). The inflight manager will ensure
+  we only handle the message once (as we should); possible duplicate
+  messages caused by retransmisions should get trapped and not
+  forwarded to the user defined publish handler.
+
 ## 0.7.0 - 2018-08-09
 
 ### Changed